ripcurrent/requirements.txt
discord.py==2.3.2
pyarrow==15.0.0
aiohttp==3.9.1
python-dotenv==1.0.0
